requirements.txt
pyppeteer==0.2.5
python-dotenv==0.18.0
openai==0.10.1
